Renames "problem operator base" files to "problem operator".
